+++ CHANGELOG.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2021-12-14
+## [Unreleased] - 2021-12-16
 
 ### Added
 * Generational evolutionary algorithms, including the following features and functionality:
   * The following EA variations:
-    * Mutation-only
+    * Standard generational model, parents replaced by children, crossover and mutation.
+    * Generational with mutation-only (no crossover).
+    * Generational, parents replaced by children, with mutually exclusive crossover and mutation (i.e.,
+      no child is result of both crossover and mutation).
+  * Crossover features (mirrors the existing features of mutation operators):
+    * CrossoverOperator interface for defining custom crossover operators.
+    * Support for hybrid crossover operators (e.g., picking randomly from set of crossover operators).
   * The following selection operators:
     * FitnessProportionalSelection
     * StochasticUniversalSampling